+++ src/playground-code-editor.ts
@@ -189,6 +189,7 @@ export class PlaygroundCodeEditor extends LitElement {
             this._valueChangingFromOutside = false;
             break;
           case 'lineNumbers':
+            this._enableOrDisableAriaLineNumberObserver();
             cm.setOption('lineNumbers', this.lineNumbers);
             break;
           case 'type':
@@ -256,6 +257,7 @@ export class PlaygroundCodeEditor extends LitElement {
     const cm = CodeMirror(
       (dom) => {
         this._cmDom = dom;
+        this._enableOrDisableAriaLineNumberObserver();
         this._resizing = true;
         requestAnimationFrame(() => {
           requestAnimationFrame(() => {
@@ -294,6 +296,43 @@ export class PlaygroundCodeEditor extends LitElement {
       }
     });
     this._codemirror = cm;
+  }
+
+  private _ariaLineNumberObserver?: MutationObserver;
+
+  /**
+   * Prevent screen readers from voicing line numbers.
+   *
+   * When line numbers are active, watch for lines inserted into the DOM by
+   * CodeMirror, and add the "aria-hidden" attribute to their line numbers.
+   *
+   * See https://github.com/codemirror/CodeMirror/issues/6578
+   */
+  private _enableOrDisableAriaLineNumberObserver() {
+    if (this.lineNumbers && !this._ariaLineNumberObserver) {
+      // Start observing newly added lines.
+      const linesParent = this._cmDom?.querySelector('.CodeMirror-code');
+      if (!linesParent) {
+        console.error('Internal playground error: .CodeMirror-code missing');
+        return;
+      }
+      this._ariaLineNumberObserver = new MutationObserver((mutations) => {
+        for (const mutation of mutations) {
+          for (const node of mutation.addedNodes) {
+            // Could be e.g. a text node. Cast so we can check for presence of
+            // querySelector with optional chaining instead of a typeof test.
+            (node as Partial<Element>)
+              .querySelector?.('.CodeMirror-gutter-wrapper')
+              ?.setAttribute('aria-hidden', 'true');
+          }
+        }
+      });
+      this._ariaLineNumberObserver.observe(linesParent, {childList: true});
+    } else if (!this.lineNumbers && this._ariaLineNumberObserver) {
+      // Line numbers are no longer rendering.
+      this._ariaLineNumberObserver.disconnect();
+      this._ariaLineNumberObserver = undefined;
+    }
   }
 
   /**